Minor refactoring and reorg around Timeline.
