+++ Packages/Mastodon/Sources/Mastodon/Timeline.swift
@@ -1,4 +1,5 @@
 import Blueprint
+import Everything
 import Foundation
 import SwiftUI
 
@@ -8,8 +9,9 @@ public enum TimelineType: Codable, Hashable, Sendable {
     case local
     case hashtag(String)
     case home
-    case list(String) // List.ID
+    case list(Mastodon.List.ID)
 
+    @available(*, deprecated, message: "Use MastodonAPI instead.")
     public var path: URLPath? {
         switch self {
         case .public:
@@ -74,6 +76,12 @@ public struct Timeline: Codable, Hashable, Sendable {
         self.timelineType = timelineType
     }
 
+    public typealias Content = PagedContent<Fetch<Element>>
+    public typealias Page = Content.Page
+    public typealias Element = Status
+}
+
+extension Timeline {
     @RequestBuilder
     func request(baseURL: URL, token: Token) -> some Request {
         switch timelineType {
@@ -87,8 +95,95 @@ public struct Timeline: Codable, Hashable, Sendable {
             MastodonAPI.Timelimes.Hashtag(baseURL: baseURL, token: token, hashtag: hashtag)
         case .home:
             MastodonAPI.Timelimes.Home(baseURL: baseURL, token: token)
-        case .list(let string):
-            MastodonAPI.Timelimes.List(baseURL: baseURL, token: token, id: Mastodon.List.ID(string)) // TODO:
+        case .list(let id):
+            MastodonAPI.Timelimes.List(baseURL: baseURL, token: token, id: id)
         }
+    }
+}
+
+public extension Service {
+    func fetchPageForTimeline(_ timeline: Timeline) async throws -> Timeline.Page {
+        let request = timeline.request(baseURL: baseURL, token: authorization.token!)
+        let fetch = Fetch(Status.self, service: self, request: request)
+        let page = try await fetch()
+        return page
+    }
+
+    func fetchTimeline(_ timeline: Timeline) async throws -> Timeline.Content {
+        fatalError()
+    }
+
+    func updateTimeline(_ timeline: Timeline) async throws -> Timeline.Content {
+        fatalError()
+    }
+}
+
+// MARK: -
+
+public struct Fetch <Element>: FetchProtocol where Element: Codable & Identifiable & Sendable, Element.ID: Comparable & Sendable {
+    let service: Service?
+    let request: (any Request)?
+
+    public init(_ elementType: Element.Type, service: Service? = nil, request: (any Request)?) {
+        self.service = service
+        self.request = request
+    }
+
+    public func callAsFunction() async throws -> Page<Self> {
+        guard let service, let request else {
+            // TODO: Log this.
+            return Page(previous: nil, next: nil, elements: [])
+        }
+        let response = PageResponse<Element> { url in
+            let request = OverrideURLRequest(content: request, overrideURL: url)
+            let fetch = Fetch(Element.self, service: service, request: request)
+            return fetch
+        }
+        let page = try await service.perform(request: request, response: response)
+        return page
+    }
+}
+
+// MARK: -
+
+struct PageResponse <Element>: Response where Element: Codable & Identifiable, Element.ID: Comparable {
+    typealias Result = Page<Fetch<Element>>
+
+    let fetchForURL: (URL) -> Fetch<Element>?
+
+    var response: some Response {
+        IfStatus(200) { data, response -> Result in
+            let elements = try JSONDecoder.mastodonDecoder.decode([Element].self, from: data)
+            var previous: Fetch<Element>?
+            var next: Fetch<Element>?
+            if let link = response.allHeaderFields["Link"] {
+                // swiftlint:disable:next force_cast
+                let link = link as! String
+                let links = try processLinks(string: link)
+                if let url = links["prev"] {
+                    previous = fetchForURL(url)
+                }
+                if let url = links["next"] {
+                    next = fetchForURL(url)
+                }
+            }
+            return Page(previous: previous, next: next, elements: elements)
+        }
+    }
+}
+
+extension Fetch: Codable {
+    public init(from decoder: Decoder) throws {
+        // TODO: Only semi codable.
+        let container = try decoder.singleValueContainer()
+        guard let service = decoder.userInfo[CodingUserInfoKey(rawValue: "service")!] as? Service else {
+            fatalError("No service set on decoder userinfo.")
+        }
+        self.service = service
+        self.request = nil
+    }
+
+    public func encode(to encoder: Encoder) throws {
+        // TODO: Only semi codable.
     }
 }
